+++ changelog.md
@@ -1,6 +1,15 @@
 ### Fabric
-- fix bug, when ServerPlayer in threadLocal is null - [**#34**](https://github.com/btwonion/magnetic/issues/34)
+- fix the bug when ServerPlayer in threadLocal is null - [**#34**](https://github.com/btwonion/magnetic/issues/34)
 - add compatibility for RightClickHarvest by jamalam
 - fix icon not loading in ModMenu
 - improve code that potentially could have changed the game in ways it shouldn't have
 - add support for 1.21.6
+- add the following breakchained blocks:
+  - BaseRailBlock
+  - DiodeBlock
+  - FallingBlockEntities
+  - LeavesBlock
+  - RedStoneWireBlock
+  - SnowLayerBlock
+  - SpongeBlock
+  - CandleCakeBlock